snapshot: builds but...wut?
