+++ Sources/ImageEmbosser/ImageEmbosser.swift
@@ -1,0 +1,9 @@
+import ArgumentParser
+
+@main
+struct RouteGuide: AsyncParsableCommand {
+  static let configuration = CommandConfiguration(
+    commandName: "image-embosser",
+    subcommands: [Serve.self]
+  )
+}
