+++ Sources/Protos/image_embosser/image_embosser.proto
@@ -1,5 +1,6 @@
 syntax = "proto3";
 
+package image_embosser;
 option go_package = "github.com/sfomuseum/go-image-emboss/grpc";
 
 service ImageEmbosser {
